* Added a public view for published datasets. * Added CleanerVersion plugin to control versions of published dataset.
NOTE: For versioning, things are easy for searches (keyword based and QBE tasks) as each search is recorded into the DB with keyword and returned JSON, so for each new revision of the dataset, we just need to collect a set of searched records as of that moment. But for survey tasks, because users may change their opinions by time after they know some new facts. And it will be more challenging for space-time annotation tasks, because each drawn annotation should be considered as a new entity, and modifications of space-time annotations are often and cannot be forbidden to have better quality of annotations.